requirements.txt
tavily-python
python-dotenv
ipykernel
langchain
langchain-openai
langchain-core
langgraph
pydantic
openai
requests
langchain_chroma
langchain-community
langchain_groq